mysite/requirements.txt
amqp==2.5.0
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-grpc==0.1.8
djangorestframework==3.10.1
djangorestframework-xml==1.4.0
elasticsearch==7.0.2
entrypoints==0.3
flake8==3.7.8
grpcio==1.22.0
grpcio-tools==1.22.0
idna==2.8
kombu==4.6.3
mccabe==0.6.1
Pillow==6.1.0
protobuf==3.9.0
psycopg2==2.8.3
pycodestyle==2.5.0
pydevd==1.6.1
pyflakes==2.1.1
pymongo==3.8.0
python-memcached==1.59
pytz==2019.1
redis==3.2.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0